Add the ability to save macros and copy them as commands (#1047)
With this PR, you can now record multiple macros by saving them and then
play them by referencing their name. You can list all of these saved
macros with 'macro list'. The saved macros are not persisted across
Talon sessions (this may cause confusion as 'save' is in the name of the
command, open to suggestions). To save a macro persistently, you can
copy the last macro or copy a specific macro by name and then paste that
in a talon file.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Stefan Schneider-Kennedy <code@stefansk.name>